requirements.txt
discord.py~=2.5.2
loguru~=0.7.3
requests~=2.32.3
pillow~=11.1.0
aiohttp~=3.11.13
httpx~=0.28.1
pydub